# apps/api/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/guest/start": {
      "post": {
        "operationId": "GuestController_start",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Guest"
        ]
      }
    },
    "/guest/join/{shareToken}": {
      "post": {
        "operationId": "GuestController_join",
        "parameters": [
          {
            "name": "shareToken",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Guest"
        ]
      }
    },
...
  },
  "info": {
    "title": "Forky API",
    "description": "Forky API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}